requirements.txt
fastapi
uvicorn
pillow
numpy
onnxruntime
opencv-python-headless